MP2.0 added, minor updates to HTMP1.1
MP2.0 - added for new revival on 24V system. Uses the board from HTMP1.0 - but is otherwise identical to MP2.1 (since MP2.2 currently is expanded dimensionally). It's configuration was basically copied from MP2.1, with changes to board, PID (copy of old MP2.0 settings), and mesh size.
HTMP1.1 - Changes back to TMC2209 since we have the Super Whopper and that works like a charm.