G101 now parses U value for singular offset
The offset is not saved to any location and is cumulative over commands. G101 U10 followed by G101 U-10 would result in returning to the same position rather than moving 20mm total.

G101 currently only takes the U flag for Z0 motor.

Not sure about the speed of this implementation over lots of commands, but may be offset by cleaning up how the multi-axis is controlled.

Current implementation is:
1. Perform the action to adjust the Z motors independently
2. Perform normal G1 plotting