Changed to save cumulative deltas (broken)
This does not actually work as intended. This commit is being saved as a checkpoint and for documentation only.  The original understanding of this setup was that doing the blocking moves to a z location would result in offsetting that axis and then all the others moving on their own. This is not the observed behavior. Instead, the current z location is updated to wherever the blocking move was to rather than retaining its old position, resulting in movement that does not reach desired positions on axes that do not have a delta of zero.