Missing axis problems, driver test config
Created a driver testing config which only has the X axis enabled, but is intended for the test bench. It also allows for pin debugging.
In the process of creating this, I discovered that my changes to add the z_deltas (which I don't think are even actively used on the ZVW version that we are currently running for non-planar) had some issues if you don't have the Z axis. Furthermore, the set_current functionality breaks if you don't have an X, Y, or Z axis so I updated that as well. Now, less things get mad if you don't have an axis.